requirements.txt
colorama==0.4.6
numpy==2.4.3
pillow==12.1.1
